chore: cleanup crate structure
