+++ .github/workflows/linux.yml
@@ -11,45 +11,26 @@ jobs:
     name: linux
 
     steps:
-    - name: checkout
-      uses: actions/checkout@v2
+    - name: Git Checkout
+      uses: actions/checkout@v4
 
-    - name: setup
-      uses: actions-rs/toolchain@v1
+    - name: Setup Rust
+      uses: dtolnay/rust-toolchain@stable
       with:
-        toolchain: nightly-x86_64-unknown-linux-gnu
+        toolchain: 1.82
         target: x86_64-unknown-linux-gnu
-        override: true
-        profile: minimal
-        default: true
+        components: clippy
 
-    - name: prep test tools
+    - name: Run Tests
       run: |
-        cargo install cargo2junit
-        mkdir -p .testresults/unit
-        mkdir -p .coverage/unit
+        cargo test
 
-    - name: generate code coverage
-      uses: actions-rs/tarpaulin@v0.1
-      with:
-        args: '-o Xml --exclude-files *_test.rs,main.rs --output-dir .coverage/unit'
-
-    - name: generate junit unit test results
+    - name: Clippy Checks
       run: |
-        cargo test -- -Z unstable-options --format json --report-time | cargo2junit > .testresults/unit/junit.xml
+        rustup component add clippy
+        cargo clippy --workspace --tests --all-features -- -D warnings
 
-    - name: setup testspace publisher
-      uses: testspace-com/setup-testspace@v1
-      with:
-        domain: ${{github.repository_owner}}
-      if: always()
-
-    - name: publish unit test results
-      run: testspace .testresults/unit/junit.xml
-      if: always()
-
-    - name: publish code coverage results
-      uses: codecov/codecov-action@v2
-      with:
-        files: .coverage/unit/cobertura.xml
-
+    - name: Check for Unused Dependencies
+      run: |
+        cargo install cargo-machete@0.7.0
+        cargo machete
