+++ .rusty-hook.toml
@@ -1,5 +1,5 @@
 [hooks]
-pre-commit = "cargo test --bin rusty-hook"
+pre-commit = "cargo clippy --workspace -- -D warnings && cargo test --workspace"
 pre-push = "cargo fmt -- --check"
 
 [logging]
