--- .vscode/cSpell.json
@@ -1,49 +1,0 @@
-{
-  "version": "0.1",
-  "language": "en",
-  "words": [
-    "applypatch",
-    "choco",
-    "chocolatey",
-    "cinst",
-    "clippy",
-    "calebcartwright",
-    "cobertura",
-    "codebases",
-    "codecov",
-    "cratesio",
-    "dependabot",
-    "devops",
-    "editorconfig",
-    "gcda",
-    "gcno",
-    "getopts",
-    "gitattributes",
-    "githook",
-    "junit",
-    "kcov",
-    "nias",
-    "pipefail",
-    "repo",
-    "rustc",
-    "rustfmt",
-    "rustup",
-    "semver",
-    "sendemail",
-    "setvariable",
-    "shellcheck",
-    "shiba",
-    "shunit",
-    "swellaby",
-    "testresults",
-    "toml",
-    "userprofile",
-    "vctools",
-    "vendored",
-    "visualstudio",
-    "vswhere"
-  ],
-  "flagWords": [
-
-  ]
-}
