--- .vscode/extensions.json
@@ -1,9 +1,0 @@
-{
-  "recommendations": [
-    "swellaby.rust-pack",
-    "swellaby.common-pack",
-    "swellaby.vscode-rust-test-adapter",
-    "hbenl.vscode-test-explorer",
-    "ms-azure-devops.azure-pipelines"
-  ]
-}
