--- .vscode/settings.json
@@ -1,48 +1,0 @@
-{
-  "files.exclude": {
-    "**/target/": true,
-    "**/*.rs.bk": true,
-    ".coverage": true,
-    ".testresults": true,
-    "**/kcov*/": true,
-    "**/*.gcda": true,
-    "**/*.gcno": true,
-    "**/cobertura.xml": true,
-    "**/tarpaulin-report.html": true
-  },
-  "cSpell.enabled": true,
-  "cSpell.ignorePaths": [
-    "Cargo.lock",
-    "**/.git/**",
-    ".vscode",
-  ],
-  "cSpell.enabledLanguageIds": [
-    "asciidoc",
-    "azure-pipelines",
-    "html",
-    "json",
-    "markdown",
-    "plaintext",
-    "powershell",
-    "properties",
-    "restructuredtext",
-    "rust",
-    "shellscript",
-    "text",
-    "toml",
-    "yaml"
-  ],
-  "cSpell.dictionaries": [
-    "companies",
-    "cpp",
-    "en_us",
-    "filetypes",
-    "misc",
-    "powershell",
-    "rust",
-    "softwareTerms"
-  ],
-  "files.associations": {
-    "**/.azure-pipelines/**": "azure-pipelines"
-  }
-}
