+++ rust-toolchain.toml
@@ -1,0 +1,2 @@
+[toolchain]
+channel = "1.82"
