+++ src/hooks/files/cli.sh
